+++ app/devices/forms.py
@@ -11,19 +11,30 @@ class DeviceRegistrationForm(FlaskForm):
     """WTForm used to register devices."""
     nom = wtforms.StringField(_l("Nom (optionnel)"),
                               validators=[Optional(), Length(max=64)])
-    mac = wtforms.StringField(_l("Adresse MAC*"),
-                              validators=[DataRequired(), MacAddress()])
     type = wtforms.StringField(_l("Type (optionnel)"),
                                validators=[Optional(), Length(max=64)])
+    mac = wtforms.StringField(_l("Adresse MAC*"),
+                              validators=[DataRequired(), MacAddress()])
     submit = wtforms.SubmitField(_l("Enregistrer l'appareil"))
+
+
+class DeviceModificationForm(FlaskForm):
+    """WTForm used to modify a device."""
+    nom = wtforms.StringField(_l("Nom (optionnel)"),
+                              validators=[Optional(), Length(max=64)])
+    type = wtforms.StringField(_l("Type (optionnel)"),
+                               validators=[Optional(), Length(max=64)])
+    mac = wtforms.StringField(_l("Adresse MAC"),
+                              validators=[DataRequired(), MacAddress()])
+    submit = wtforms.SubmitField(_l("Modifier l'appareil"))
 
 
 class DeviceTransferForm(FlaskForm):
     """WTForm used to transfer a device."""
     nom = wtforms.StringField(_l("Nom (optionnel)"),
                               validators=[Optional(), Length(max=64)])
-    mac = wtforms.StringField(_l("Adresse MAC"),
-                              validators=[DataRequired(), MacAddress()])
     type = wtforms.StringField(_l("Type (optionnel)"),
                                validators=[Optional(), Length(max=64)])
+    mac = wtforms.StringField(_l("Adresse MAC"),
+                              validators=[DataRequired(), MacAddress()])
     submit = wtforms.SubmitField(_l("Transférer l'appareil"))