+++ app/rooms/forms.py
@@ -20,6 +20,15 @@ class RentalRegistrationForm(FlaskForm):
     submit = wtforms.SubmitField(_l("Enregistrer"))
 
 
+class RentalModificationForm(FlaskForm):
+    """WTForm used to modify a room rental."""
+    start = html5.DateField(_l("Début de la location"),
+                            validators=[DataRequired(), PastDate()])
+    end = html5.DateField(_l("Fin de la location (optionnel)"),
+                          validators=[Optional(), FutureDate()])
+    submit = wtforms.SubmitField(_l("Modifier"))
+
+
 class RentalTransferForm(FlaskForm):
     """WTForm used to terminate a room rental, before creating a new one."""
     end = html5.DateField(_l("Fin de la location"),